% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_lt_category.R
\name{adjust_lt_category}
\alias{adjust_lt_category}
\title{Adjust life table category (universal, location-specific, outlier) based on provided designations.
Designations may come from machine vision or manual review.}
\usage{
adjust_lt_category(dt, review, id_vars, loc_list)
}
\arguments{
\item{dt}{data.table with empirical life tables}

\item{review}{data.table with id variables, plus life_table_category_id representing desired category}

\item{id_vars}{vector of id variables, not including age}

\item{loc_list}{vector of location_ids for which we will add back in excluded ELTs as loc-specific and smooth width 7}
}
\value{
data.table with modified life table category id
}
\description{
Adjust life table category (universal, location-specific, outlier) based on provided designations.
Designations may come from machine vision or manual review.
}
