% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_age_split_deaths.R
\name{agg_age_split_deaths}
\alias{agg_age_split_deaths}
\title{Aggregate age-split empirical deaths. For extension of ELTs we only want to use pre-split data.}
\usage{
agg_age_split_deaths(dt, unsplit_terminal_ages)
}
\arguments{
\item{dt}{data.table with columns ihme_loc_id, year, sex, source_type, age, age_group_name, deaths}

\item{unsplit_terminal_ages}{data.table specifying the terminal ages of pre-age-split data with columns ihme_loc_id, year, sex, source_type, terminal_age}
}
\value{
returns dt but with any age-split data as specified in unsplit_terminal_ages re-aggregated
}
\description{
Aggregate age-split empirical deaths. For extension of ELTs we only want to use pre-split data.
}
