% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade_select_lts.R
\name{cascade_select_lts}
\alias{cascade_select_lts}
\title{Calculate thresholds for which to include lifetables in the universal or location-specific pool
After segmenting lifetables, keep each country/year/sex observation with the least smoothing, by lifetable category}
\usage{
cascade_select_lts(empir_lt, id_vars, hiv_cdr, group_1_list, pop_age_pools,
  country_pop, loc_spec_comp, loc_spec_manual, outliers_5q0_45q15,
  outliers_comp, outliers_manual)
}
\arguments{
\item{empir_lt}{data.table with variables ihme_loc_id, sex, source_type, deaths_nid, deaths_underlying_nid, smooth_width, age, qx}

\item{id_vars}{character vector of variables that, in addition to age and smooth_width, uniquely identify all observations}

\item{hiv_cdr}{data.table of adult HIV CDR, coming from best 45q15 estimate run. ihme_loc_id, year, sex}

\item{group_1_list}{character vector of ihme_loc_ids that are Group 1A or Group 1B}

\item{pop_age_pools}{data.table with ihme_loc_id and pop_pool_age, from population old-age pooling methods}

\item{country_pop}{data.table with ihme_loc_id, year_id, mean_pop (representing population). Used for conditional population-based outliering/cutoffs}

\item{loc_spec_comp}{data.table with ihme_loc_id, sex, year, source_type, comp_under_85. Countries that violate minimum completeness bounds and need to be forced to location-specific life-tables}

\item{loc_spec_manual}{data.table with ihme_loc_id, sex, year, source_type, smooth_width, manual_loc_specific. Where manual_loc_specific=1 we'll set to location-specific.}

\item{outliers_5q0_45q15}{data.table with ihme_loc_id, sex, year, source_type, child_adult. Where child_adult=1 we'll set to outlier all smooth widths.}

\item{outliers_comp}{data.table with ihme_loc_id, sex, year, source_type, comp_under_50. Where comp_under_50=1 we'll outlier all smooth widths.}

\item{outliers_manual}{data.table with ihme_loc_id, sex, year, source_type, smooth_width, manual_outlier. Where manual_outlier=1 we'll outlier.}
}
\value{
returns empir_lt with the same variables
}
\description{
Calculate thresholds for which to include lifetables in the universal or location-specific pool
After segmenting lifetables, keep each country/year/sex observation with the least smoothing, by lifetable category
}
