% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_ax.R
\name{extend_ax}
\alias{extend_ax}
\title{Extend ax values for ages 75-105 using HMD regression parameters}
\usage{
extend_ax(empir_lt, hmd_ax_results)
}
\arguments{
\item{empir_lt}{data.table with columns: ihme_loc_id, sex, year, age (numeric), mx, ax}

\item{hmd_ax_results}{data.table with variables par_mx, par_smx, par_con, sex, age}
}
\value{
returns empir_lt with the same variables, but modified values for ax for ages 75 and above
}
\description{
Extend ax values for ages 75-105 using HMD regression parameters
}
