% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_qx_kt.R
\name{extend_qx_kt}
\alias{extend_qx_kt}
\title{Extend qx values up to age group 105-109 using Kannisto-Thatcher model}
\usage{
extend_qx_kt(data, byvars, age_width = 20)
}
\arguments{
\item{data}{data.table with columns: ihme_loc_id, sex, year, age (numeric), mx, qx, ax}

\item{byvars}{character vector containing the variable names of all variables that uniquely identify the observations (except for age)}

\item{age_width}{numeric, default = 20, inclusive width of interval of data to include in years.
example: data that goes to age 90-94 would be fit starting with data from 70-74 when age_width=20.}
}
\value{
returns data.table with qx and ax extended to age group 105-109

 Initial parameter values selected using methodology from
 Feehan, D.M. Demography (2018) 55: 2025. https://doi.org/10.1007/s13524-018-0728-x

 Additional reference: HMD methods
 https://www.mortality.org/Public/Docs/MethodsProtocol.pdf
}
\description{
Extend qx values up to age group 105-109 using Kannisto-Thatcher model
}
