% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_qx_kt.R
\name{fit.kt}
\alias{fit.kt}
\title{Fit K-T model (one life table)}
\usage{
fit.kt(data, age_width = 20)
}
\arguments{
\item{data}{data.table with columns: ihme_loc_id, sex, year, age (numeric), mx, qx}

\item{age_width}{numeric, default = 20, inclusive width of interval of data to include in years.
example: data that goes to age 90-94 would be fit starting with data from 70-74 when age_width=20.}
}
\value{
returns vector of K-T parameters (a, b)
}
\description{
Fit K-T model (one life table)
}
