% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_qx_kt.R
\name{fit.predict.kt}
\alias{fit.predict.kt}
\title{Fit Kannisto-Thatcher model and predict (one life table)}
\usage{
fit.predict.kt(data, byvars, age_width = 20)
}
\arguments{
\item{data}{data.table with columns: ihme_loc_id, sex, year, age (numeric), mx, qx}

\item{byvars}{character vector containing the variable names of all variables that uniquely identify the observations (except for age)}

\item{age_width}{numeric, default = 20, inclusive width of interval of data to include in years.
example: data that goes to age 90-94 would be fit starting with data from 70-74 when age_width=20.}
}
\value{
returns data.table with qx extended to age group 105-109 for one life table
}
\description{
Fit Kannisto-Thatcher model and predict (one life table)
}
