% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_qx.R
\name{gen_80plus_ax}
\alias{gen_80plus_ax}
\title{Iterate terminal qx & lx values to minimize the difference to observed terminal mx
TODO: right now this only works for terminal age 95+}
\usage{
gen_80plus_ax(dt, ax_params)
}
\arguments{
\item{dt}{data.table with variables: ihme_loc_id, year, source_name, sex, age, age_length, dx, ax, qx, terminal_age_start, mx_term}

\item{ax_params}{parameters for ax extension from mlt}

\item{id_vars}{character vector of id variables that uniquely identify each observation (last one must be age)}

\item{n_iterations}{numeric, number of iterations to run

Note: ltcore required}
}
\value{
None. Modifies given data.table in-place
}
\description{
Iterate terminal qx & lx values to minimize the difference to observed terminal mx
TODO: right now this only works for terminal age 95+
}
