% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_initial_ax.R
\name{gen_initial_ax}
\alias{gen_initial_ax}
\title{Generate Starting ax Values for all ages}
\usage{
gen_initial_ax(dt, id_vars)
}
\arguments{
\item{dt}{data.table with variables: ihme_loc_id, sex, year, age, mx}
}
\value{
None. Modifies given data.table in-place
}
\description{
Initiate ax at midpoint for most ages, but for under-5 do a custom ax calculation based on _______
}
