% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_age_split_deaths.R
\name{get_unsplit_terminal_age}
\alias{get_unsplit_terminal_age}
\title{Get terminal age for life tables from empirical deaths output}
\usage{
get_unsplit_terminal_age(death_number_empirical_version)
}
\arguments{
\item{death_number_empirical_version}{integer, version_id for empirical deaths}
}
\value{
returns data.table with columns ihme_loc_id, year, sex, source_type, terminal_age
}
\description{
Get terminal age for life tables from empirical deaths output
}
