% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hmd_lts.R
\name{import_hmd_lts}
\alias{import_hmd_lts}
\title{Import all HMD lifetables, format them long, and reunite East and West Germany for HMD}
\usage{
import_hmd_lts(lt_pars = c("qx"))
}
\arguments{
\item{lt_pars}{list of strings, names of life table parameters to be included in output (i.e. qx, ax, etc)}
}
\value{
data.table with variables ihme_loc_id, sex, age (numeric, values 0, 1, 5 (5) 110), source (character), qx
}
\description{
Import all HMD lifetables, format them long, and reunite East and West Germany for HMD
}
