% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_ax.R
\name{iterate_ax}
\alias{iterate_ax}
\title{Generate ax values by iterating ax values to minimize the difference across ax, dx, etc.}
\usage{
iterate_ax(dt, id_vars, n_iterations)
}
\arguments{
\item{dt}{data.table with variables: ihme_loc_id, year, source_name, sex, age, age_length, dx, ax, qx}

\item{id_vars}{character vector of id variables that uniquely identify each observation (last one must be age)}

\item{n_iterations}{numeric, number of iterations to run}
}
\value{
None. Modifies given data.table in-place
}
\description{
See Preston section 3.2 (pg 45) on strategies for choosing ax values
With this method, we can estimate ax from dx, but we iterate through dx because it changes when we change ax, so there is a circular relationship.
}
