% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_qx_kt.R
\name{kt.ax}
\alias{kt.ax}
\title{Kannisto-Thatcher model, in 5ax space
See pg 69 Preston Demography textbook for equation for nax from mu}
\usage{
kt.ax(x, params)
}
\arguments{
\item{x}{age (numeric)}

\item{params}{vector of K-T parameters (a, b)}
}
\value{
returns 5ax output of K-T model for age x and parameters a, b (numeric)
}
\description{
Kannisto-Thatcher model, in 5ax space
See pg 69 Preston Demography textbook for equation for nax from mu
}
