% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_qx_kt.R
\name{kt.mu}
\alias{kt.mu}
\title{Kannisto-Thatcher model, in force of mortality (mu) space
mux = [a * exp(bx)] / [1 + a * exp(bx)]}
\usage{
kt.mu(x, params)
}
\arguments{
\item{x}{age (numeric)}

\item{params}{vector of K-T parameters (a, b)}
}
\value{
returns mux output of K-T model for age x and parameters a, b (numeric)
}
\description{
Kannisto-Thatcher model, in force of mortality (mu) space
mux = [a * exp(bx)] / [1 + a * exp(bx)]
}
