% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_weighted_average.R
\name{moving_weighted_average}
\alias{moving_weighted_average}
\title{Generate a moving weighted average}
\usage{
moving_weighted_average(x, weights)
}
\arguments{
\item{x}{vector of data to be smoothed}

\item{weights:}{a list of weights to use in smoothing}
}
\value{
returns vector of smoothed data
}
\description{
Generate a moving weighted average
}
