% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_data.R
\name{outlier_data}
\alias{outlier_data}
\title{Outliers from 5q0 and 45q15 modeling, return data.table of outliers to be used by cascade_select_lts}
\usage{
outlier_data(run_id_5q0_estimate, run_id_45q15_estimate, gbd_year)
}
\arguments{
\item{run_id_5q0_estimate}{integer, run ID of 5q0 estimates to pull underlying data and outliers from}

\item{run_id_45q15_estimate}{integer, run ID of 5q0 estimates to pull underlying data and pull outliers from}

\item{gbd_year}{integer, gbd year (ex: 2017, 2019, 2020, etc.)}
}
\value{
data.table with ihme_loc_id, year, sex, source_type, child_adult=1, based on 5q0 and 45q15 outliers
}
\description{
Outliers from 5q0 and 45q15 modeling, return data.table of outliers to be used by cascade_select_lts
}
