% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_qx_kt.R
\name{pred.kt}
\alias{pred.kt}
\title{Expand ages and predict K-T (one life table)}
\usage{
pred.kt(data, params, byvars)
}
\arguments{
\item{data}{data.table with columns: ihme_loc_id, sex, year, age (numeric), mx, qx}

\item{params}{vector of parameters for K-T model (a, b)}

\item{byvars}{character vector containing the variable names of all variables that uniquely identify the observations (except for age)}
}
\value{
returns data.table with qx extended to age group 105-109 for one life table
}
\description{
Expand ages and predict K-T (one life table)
}
