% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_lts.R
\name{smooth_lts}
\alias{smooth_lts}
\title{Smooth targeted lifetable qx values by using lifetables in the same country/sex years surrounding it.}
\usage{
smooth_lts(empir_lt, by_vars, moving_average_weights)
}
\arguments{
\item{empir_lt}{data.table with columns: ihme_loc_id, sex, year, source, age (numeric), qx, lx}

\item{by_vars}{character vector of variables that would uniquely identify each observation if combined with the variable year.}

\item{moving_average_weights}{numeric vector of weights to use for moving average (if using). Must be symmetric, and will produce NA values for places where there are not enough values on either side
Example: c(1, 1.5, 6, 1.5, 1) uses a relative weight of 6 for the observation itself, 1.5 for lag/lead 1, 1 for lag/lead 2, and 0 for all other observations}

\item{smoothing_targets}{data.table with variables ihme_loc_id, sex}
}
\value{
returns empir_lt with the same variables, but modified values for qx and lx based off of the combined, smoothed lifetables
}
\description{
Smooth targeted lifetable qx values by using lifetables in the same country/sex years surrounding it.
}
