% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_qx_kt.R
\name{sse}
\alias{sse}
\title{Optimization function for K-T model
SSE = sum of (qx - qx_pred)^2}
\usage{
sse(params, data)
}
\arguments{
\item{params}{vector of K-T parameters (a, b)}

\item{data}{data.table with columns: ihme_loc_id, sex, year, age (numeric), mx, qx}
}
\value{
returns sum of square errors (numeric)
}
\description{
Optimization function for K-T model
SSE = sum of (qx - qx_pred)^2
}
