% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_qx_kt.R
\name{start.kt}
\alias{start.kt}
\title{Calculate starting parameters a and b for input to optim (one life table)
Based on logit(mx) ~ ln(a) + b*age}
\usage{
\method{start}{kt}(data)
}
\arguments{
\item{data}{data.table with columns: ihme_loc_id, sex, year, age (numeric), mx, qx}
}
\value{
returns list of K-T parameters (a, b)
}
\description{
Calculate starting parameters a and b for input to optim (one life table)
Based on logit(mx) ~ ln(a) + b*age
}
